1. 消费金额限制,金额保存32位整数 2. 正在消费，收到app开始消费返回失败，返回请求的订单号 3. 结束消费命令区别订单号 4. SIM卡唯一号绑定程序

+++ src/hardware/digitron.h
@@ -18,7 +18,7 @@ void setOnFlag(void);
 void setOffFlag(void);
 
 //数码管显示金额
-void display(u16 money);
+void display(u32 money);
 
 
 #endif
